requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
eventlet==0.33.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
